refactor: apply review feedback
