+++ apps/api/src/lessons/lessons.module.ts
@@ -4,7 +4,7 @@ import { S3Service } from "src/file/s3.service";
 
 import { AdminLessonItemsService } from "./adminLessonItems.service";
 import { AdminLessonsService } from "./adminLessons.service";
-import { LessonsController } from "./api/lessons.controller";
+import { LessonsController } from "./lessons.controller";
 import { LessonsService } from "./lessons.service";
 import { AdminLessonItemsRepository } from "./repositories/adminLessonItems.repository";
 import { AdminLessonsRepository } from "./repositories/adminLessons.repository";
