+++ apps/api/src/lessons/schemas/lessonQuery.ts
@@ -1,4 +1,7 @@
-import { Type, type Static } from "@sinclair/typebox";
+import { type Static, Type } from "@sinclair/typebox";
+
+import type { UUIDType } from "src/common";
+import type { UserRole } from "src/users/schemas/user-roles";
 
 export const lessonSortFields = ["title", "createdAt", "state", "itemsCount"] as const;
 
@@ -31,3 +34,12 @@ export const lessonsFilterSchema = Type.Object({
 });
 
 export type LessonsFilterSchema = Static<typeof lessonsFilterSchema>;
+
+export type LessonsQuery = {
+  filters?: LessonsFilterSchema;
+  sort?: SortLessonFieldsOptions;
+  page?: number;
+  perPage?: number;
+  currentUserRole?: UserRole;
+  currentUserId?: UUIDType;
+};
